MGFW-12273 New class RedirectDetails and change BlockDetails to extend it for support for existing apps.
